mce-core/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
tiktoken>=0.6.0
sentence-transformers>=2.5.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
rich>=13.7.0
numpy>=1.26.0
beautifulsoup4>=4.12.0
markdownify>=0.12.0
pyyaml>=6.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.0
aiosqlite>=0.20.0
typer>=0.9.0
